swapi_api/requirements.txt
requests
graphene-django
djangorestframework
